requirements.txt
fastapi
python-dotenv
pydantic-ai
pydantic
openai